Rip out --docker-run.
