+++ README.md
@@ -14,6 +14,5 @@ Telepresence allows you to run your code locally while still:
 1. Giving your code access to Services in a remote Kubernetes cluster.
 2. Giving your code access to cloud resources like AWS RDS or Google PubSub.
 3. Allowing Kubernetes to access your code as if it were in a normal pod within the cluster.
-4. Running your local code either as a normal process or inside a local Docker container.
 
 Telepresence is part of the [Blackbird Project](https://www.datawire.io/blackbird/), and licensed under the Apache 2.0 License.
